refactor(graph_view): extract text-metrics module (#75)
Pure code move: display_width, char_width_with_vs16, truncate_to_width,
format_date_field, and the VS16 / DATE_FIELD_WIDTH consts (plus their unit
tests) move from the graph_view god module into a new metrics submodule.
graph_view.rs becomes graph_view/mod.rs; display_width is re-exported so the
external contract is unchanged. No behavior changes.
